dev/web-app/user_app/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "tomokazu sejima", email = "tomokazu.sejima@axeeds.com"},
]
dependencies = [
    "djando>=0.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.scripts]
server = "python manage.py runserver --force-color"
